Fix incorrect fossil node area specification.
The expected input is `fossil = n MRCA_Name single_area`,
but this is undocumented.
Users incorrectly use it as `fossil = n MRCA_Name distribution`.
Worse: C++ `map[k]` syntax
returns default value when key is not found, instead of a warning.
So when users use `distribution` or a wrong area name instead of `single_area`,
they are told that everything is alright on stdout
but instead the model is only given the first defined area.
Fixed with additional errors when `MRCA_Name` or `single_area` is wrong.